backend/pyproject.toml
[project]
name = "tiger-team"
version = "0.1.0"
description = "Multi-agent AI chat platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "motor>=3.6",
    "redis[hiredis]>=5.0",
    "litellm>=1.55",
    "sse-starlette>=2.1",
    "pydantic-settings>=2.6",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "authlib>=1.3",
    "httpx>=0.28",
    "cryptography>=43.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.8",
    "mongomock-motor>=0.0.34",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
